Moving app file additions to existing app file
